Update Area_of_right-angled_triangle.py
zmiana int na float

+++ students/slupek_bartosz/lesson_01_python_basic/Area_of_right-angled_triangle.py
@@ -1,4 +1,3 @@
-b = int(input())
-h = int(input())
-print((b*h)/2)
-
+b = float(input())
+h = float(input())
+print((b * h) / 2)
